working layout test with mocks
