+++ package.json
@@ -36,6 +36,7 @@
     "@vitejs/plugin-vue": "^5.0.4",
     "@vue/eslint-config-airbnb": "8.0.0",
     "@vue/eslint-config-typescript": "^13.0.0",
+    "@vue/test-utils": "^2.4.5",
     "@vue/tsconfig": "^0.5.1",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
@@ -51,6 +52,7 @@
     "msw": "^2.2.13",
     "pinia": "^2.1.7",
     "prettier": "^3.2.5",
+    "resize-observer-polyfill": "^1.5.1",
     "sass": "^1.71.1",
     "typescript": "^5.4.2",
     "unplugin-auto-import": "^0.17.5",
