+++ tests/unit/layouts/default.spec.ts
@@ -1,14 +1,10 @@
 import { describe, expect, test, vi } from "vitest";
 import {createRouter, createWebHistory} from "vue-router/auto";
-import { createTestingPinia } from "@pinia/testing";
-import { createVuetify } from 'vuetify'
-import * as components from 'vuetify/components'
-import * as directives from 'vuetify/directives'
 import {defineComponent} from "vue";
 import { render, screen, within } from "@testing-library/vue";
-import mockAppStore from "../mocks/mockAppStore";
 import Default from "@/layouts/default.vue";
-import {MOCK_APP_CONFIG} from "../mocks/mockObjects";
+import {mockVuetify} from "../mocks/mockVuetify";
+import {mockPinia} from "../mocks/mockPinia";
 
 const router = createRouter({
     history: createWebHistory(),
@@ -18,28 +14,7 @@ const router = createRouter({
     ]
 });
 
-// TODO: pull this out into a mock util
-const vuetify = createVuetify({
-    components,
-    directives,
-})
-
-// TODO: can we avoid this by using test pinia?
-//mockAppStore();
-vi.stubGlobal('ResizeObserver', vi.fn());
-vi.stubGlobal("observer.observe", vi.fn());  // this is a vuetify issue
-
-const initialState = {
-    app: {
-        loading: false,
-        selectedIndicator: "",
-        appConfig: MOCK_APP_CONFIG,
-        admin1Indicators: {},
-        admin1Geojson: {},
-        admin2Indicators: {},
-        admin2Geojson: {}
-    }
-};
+global.ResizeObserver = require('resize-observer-polyfill')
 
 const renderLayout = async() => {
     router.push("/");
@@ -48,7 +23,7 @@ const renderLayout = async() => {
     return render(Default, {
        global: {
            router,
-           plugins: [vuetify, createTestingPinia()]
+           plugins: [mockVuetify, mockPinia()]
        }
     });
 };
@@ -56,6 +31,7 @@ const renderLayout = async() => {
 describe("default layout", () => {
     test("displays app title from app config", async () => {
         await renderLayout();
+        expect(await screen.findByText("MockApp")).toBeVisible();
     });
     /*test("initialises data on load", () => {
