+++ tests/unit/mocks/mockVuetify.ts
@@ -1,0 +1,8 @@
+import { createVuetify } from "vuetify";
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+
+export const mockVuetify = createVuetify({
+    components,
+    directives,
+});
