+++ vite.config.mts
@@ -65,5 +65,14 @@ export default defineConfig({
     },
     server: {
         port: 3000
+    },
+    test: {
+        globals: true,
+        environment: 'jsdom',
+        server: {
+            deps: {
+                inline: ['vuetify'],
+            },
+        },
     }
 });
